refactor: types 파일 생성 후 interface 및 type 관련 코드 작성
